add middleware

+++ api/graph/resolver/request.resolvers.go
@@ -2,12 +2,18 @@ package resolver
 
 import (
 	"context"
+	"errors"
 	"time"
 
 	"github.com/PwrFr/gqlgen/graph/model"
+	"github.com/PwrFr/gqlgen/graph/resolver/middleware"
 )
 
-func (m *mutationResolver) CreateRequest(ctx context.Context, input model.NewRequest) (*model.RequestOutput, error) {
+func (m *mutationResolver) CreateRequest(ctx context.Context, input model.NewRequest, token *string) (*model.RequestOutput, error) {
+	if auth := middleware.RoleValid("student", token); !auth {
+		return nil, errors.New("Forbidden lol")
+	}
+
 	new_req := &model.RequestOutput{
 		RoomID:          input.RoomID,
 		RequestPurpose:  input.RequestPurpose,
@@ -22,7 +28,11 @@ func (m *mutationResolver) CreateRequest(ctx context.Context, input model.NewReq
 	return m.RepoDB.InsertRequest(new_req)
 }
 
-func (m *mutationResolver) UpdateRequest(ctx context.Context, input model.Approve) (*model.ApproveOutput, error) {
+func (m *mutationResolver) UpdateRequest(ctx context.Context, input model.Approve, token *string) (*model.ApproveOutput, error) {
+	if auth := middleware.RoleValid("staff", token); !auth {
+		return nil, errors.New("Forbidden lol")
+	}
+
 	new_req := &model.ApproveOutput{
 		RequestID:       input.RequestID,
 		RequestStatus:   input.RequestStatus,
@@ -47,6 +57,10 @@ func (r *queryResolver) RequestByID(ctx context.Context, accountID string) ([]*m
 	return r.RepoDB.GetRequestById(accountID)
 }
 
-func (m *mutationResolver) RemoveRequest(ctx context.Context, req_id int) (*string, error) {
+func (m *mutationResolver) RemoveRequest(ctx context.Context, req_id int, token *string) (*string, error) {
+	if auth := middleware.RoleValid("student", token); !auth {
+		return nil, errors.New("Forbidden lol")
+	}
+
 	return m.RepoDB.RemoveRequest(req_id)
 }
